@@ -43,10 +43,10 @@
     ]
   },
   "dependencies": {
-    "@ant-design/charts": "^1.2.13",
-    "@types/node": "^16.10.2",
-    "antd": "^4.16.13",
-    "axios": "^0.21.4",
+    "@ant-design/charts": "^1.3.3",
+    "@types/node": "^17.0.8",
+    "antd": "^4.18.2",
+    "axios": "^0.24.0",
     "classnames": "^2.3.1",
     "clipboard": "^2.0.8",
     "countup.js": "^2.0.8",
@@ -56,61 +56,61 @@
     "moment": "^2.29.1",
     "nprogress": "^0.2.0",
     "react": "^17.0.2",
-    "react-countup": "^6.0.0",
+    "react-countup": "^6.1.0",
     "react-custom-scrollbars": "^4.2.1",
     "react-document-title": "^2.0.3",
     "react-dom": "^17.0.2",
-    "react-intl": "^5.20.12",
+    "react-intl": "^5.24.1",
     "react-loadable": "^5.5.0",
-    "react-redux": "^7.2.5",
-    "react-router-dom": "^5.3.0",
+    "react-redux": "^7.2.6",
+    "react-router-dom": "^6.2.1",
     "react-transition-group": "^4.4.2",
-    "redux": "^4.1.1",
-    "redux-thunk": "^2.3.0",
-    "screenfull": "^5.1.0",
+    "redux": "^4.1.2",
+    "redux-thunk": "^2.4.1",
+    "screenfull": "^6.0.0",
     "vite-plugin-importer": "^0.2.5"
   },
   "devDependencies": {
-    "@babel/preset-env": "^7.15.6",
-    "@babel/preset-react": "^7.14.5",
-    "@babel/preset-typescript": "^7.15.0",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.1.0",
-    "@testing-library/user-event": "^13.2.1",
+    "@babel/preset-env": "^7.16.7",
+    "@babel/preset-react": "^7.16.7",
+    "@babel/preset-typescript": "^7.16.7",
+    "@testing-library/jest-dom": "^5.16.1",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
     "@types/classnames": "^2.3.1",
     "@types/clipboard": "^2.0.7",
-    "@types/jest": "^27.0.2",
-    "@types/js-cookie": "^2.2.7",
+    "@types/jest": "^27.4.0",
+    "@types/js-cookie": "^3.0.1",
     "@types/nprogress": "^0.2.0",
-    "@types/react": "^17.0.24",
+    "@types/react": "^17.0.38",
     "@types/react-countup": "^4.3.1",
-    "@types/react-custom-scrollbars": "^4.0.8",
+    "@types/react-custom-scrollbars": "^4.0.10",
     "@types/react-document-title": "^2.0.5",
-    "@types/react-dom": "^17.0.9",
+    "@types/react-dom": "^17.0.11",
     "@types/react-loadable": "^5.5.6",
-    "@types/react-redux": "^7.1.18",
-    "@types/react-router-dom": "^5.3.0",
-    "@types/react-transition-group": "^4.4.3",
-    "@typescript-eslint/eslint-plugin": "^4.31.2",
-    "@typescript-eslint/parser": "^4.31.2",
+    "@types/react-redux": "^7.1.21",
+    "@types/react-router-dom": "^5.3.2",
+    "@types/react-transition-group": "^4.4.4",
+    "@typescript-eslint/eslint-plugin": "^5.9.0",
+    "@typescript-eslint/parser": "^5.9.0",
     "@vitejs/plugin-react-refresh": "^1.3.6",
-    "babel-jest": "^27.2.2",
+    "babel-jest": "^27.4.6",
     "cross-env": "^7.0.3",
-    "eslint": "^7.32.0",
+    "eslint": "^8.6.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.26.0",
-    "husky": "^7.0.2",
-    "jest": "^27.2.2",
-    "less": "^4.1.1",
-    "lint-staged": "^11.1.2",
-    "prettier": "^2.4.1",
+    "eslint-plugin-react": "^7.28.0",
+    "husky": "^7.0.4",
+    "jest": "^27.4.6",
+    "less": "^4.1.2",
+    "lint-staged": "^12.1.5",
+    "prettier": "^2.5.1",
     "rimraf": "^3.0.2",
-    "ts-jest": "^27.0.5",
-    "ts-node": "^10.2.1",
-    "typescript": "^4.4.3",
-    "vite": "^2.5.10",
-    "vite-plugin-style-import": "^1.2.1",
-    "vite-plugin-html": "^2.1.0"
+    "ts-jest": "^27.1.2",
+    "ts-node": "^10.4.0",
+    "typescript": "^4.5.4",
+    "vite": "^2.7.10",
+    "vite-plugin-style-import": "^1.4.1",
+    "vite-plugin-html": "^2.1.2"
   }
 }
