@@ -43,7 +43,9 @@ const VBasicTree: React.FC<IVBasicTreeProps> = props => {
   generateList(treeData, flattenData)
 
   /** 模糊搜索树中的子节点 */
-  const handleSearch: ChangeEventHandler<any> = (e: ChangeEvent<HTMLInputElement>) => {
+  const handleSearch: ChangeEventHandler<any> = (
+    e: ChangeEvent<HTMLInputElement>,
+  ) => {
     const {value} = e.target
     // 获取展开的keys数组
     const expandedKeys = getExpandeKeysInDataListByValue(
