@@ -1,19 +1,25 @@
-export const noop = () => {};
+export const noop = () => {}
 
 export function on<T extends Window | Document | HTMLElement | EventTarget>(
   obj: T | null,
   ...args: Parameters<T['addEventListener']> | [string, Function | null, ...any]
 ): void {
   if (obj && obj.addEventListener) {
-    obj.addEventListener(...(args as Parameters<HTMLElement['addEventListener']>));
+    obj.addEventListener(
+      ...(args as Parameters<HTMLElement['addEventListener']>),
+    )
   }
 }
 
 export function off<T extends Window | Document | HTMLElement | EventTarget>(
   obj: T | null,
-  ...args: Parameters<T['removeEventListener']> | [string, Function | null, ...any]
+  ...args:
+    | Parameters<T['removeEventListener']>
+    | [string, Function | null, ...any]
 ): void {
   if (obj && obj.removeEventListener) {
-    obj.removeEventListener(...(args as Parameters<HTMLElement['removeEventListener']>));
+    obj.removeEventListener(
+      ...(args as Parameters<HTMLElement['removeEventListener']>),
+    )
   }
 }
