@@ -1,11 +1,11 @@
-import { useEffect, useRef } from "react";
+import {useEffect, useRef} from 'react'
 
 export function useTitle(title: string) {
-  const prevTitleRef = useRef(document.title);
+  const prevTitleRef = useRef(document.title)
   useEffect(() => {
-    document.title = title;
+    document.title = title
     return () => {
-      document.title = prevTitleRef.current;
-    };
-  }, [title]);
+      document.title = prevTitleRef.current
+    }
+  }, [title])
 }
