@@ -6,10 +6,4 @@ const NotFound = lazy(() => import('./error/404'))
 const TableCom = lazy(() => import('./component/table'))
 const FormCom = lazy(() => import('./component/form'))
 
-export {
-  Login,
-  Dashboard,
-  NotFound,
-  TableCom,
-  FormCom,
-}
+export {Login, Dashboard, NotFound, TableCom, FormCom}
